imganalize/requirements.txt
Django==6.0.3
pillow==12.1.1
pymongo==4.6.3